third_party/blink/web_tests/external/wpt/css/css-animations/animationevent-interface-expected.txt

This is a testharness.js-based test.
[FAIL] AnimationEventInit animation property sets value
  assert_equals: expected (object) object "[object CSSAnimation]" but got (undefined) undefined
Harness: the test ran to completion.
